Added support for "cmd" and "exec" in language independent way

+++ gSKI/src/gSKI_DoNotTouch.cpp
@@ -367,7 +367,7 @@ namespace gSKI
                   return dest;
                }
                determine_possible_symbol_types_for_string (sym->sc.name,
-                  strlen (sym->sc.name),
+                  (int)strlen (sym->sc.name),
                   &possible_id,
                   &possible_var,
                   &possible_sc,
